requirements.txt
aiohttp==1.0.2
aiopg==0.11.0
PyYAML==3.12

